feat(refs T35963): Move store to folder statement
