+++ client/js/bundles/procedure/administrationSplitStatement.js
@@ -12,7 +12,7 @@
  */
 
 import { initialize } from '@DpJs/InitVue'
-import SplitStatementStore from '@DpJs/store/procedure/SplitStatementStore'
+import SplitStatementStore from '@DpJs/store/statement/SplitStatementStore'
 import SplitStatementView from '@DpJs/components/statement/splitStatement/SplitStatementView'
 
 const components = {
